Support for Humio with Elasticsearch Bulk API
Added simple authentication support via providing an Authorization
header on the request for NetUtils.

This is not to be confused with Basic Auth, if you need basic auth you
need to remember to provide its configuration value as "Basic
base64encodedsecrethere".

Wiremock added as dependency to provide some stubbed expectations on the
humio responses.

Also borrowed the Proxy code from AbstractHttpPostJsonLogEventObserver,
this is however untested.

https://library.humio.com/stable/docs/ingesting-data/log-shippers/other-log-shippers/#elasticsearch-bulk-api